test(consensus): move the `IngressPayload` serialization/deserialization unit test from `test_utilities/types/` to `rs/types/types` (#3384)
It makes little sense to me to have unit tests for `IngressPayload` in a
test utilities crate

Note: to avoid cyclic dependency I'm not using `SignedIngressBuilder` to
construct a `SignedIngress`
