+++ rs/test_utilities/types/Cargo.toml
@@ -14,6 +14,5 @@ ic-types-test-utils = { path = "../../types/types_test_utils" }
 rand = { workspace = true }
 
 [dev-dependencies]
-assert_matches = { workspace = true }
 bincode = { workspace = true }
 serde_cbor = { workspace = true }
