+++ rs/types/types/src/batch/ingress.rs
@@ -204,41 +204,45 @@ mod tests {
         },
         time::expiry_time_from_now,
     };
+    use assert_matches::assert_matches;
+    use std::convert::TryFrom;
+
+    fn fake_http_call_content(method_name: &str) -> HttpCallContent {
+        HttpCallContent::Call {
+            update: HttpCanisterUpdate {
+                canister_id: Blob(vec![42; 8]),
+                method_name: method_name.to_string(),
+                arg: Blob(b"".to_vec()),
+                sender: Blob(vec![0x05]),
+                nonce: Some(Blob(vec![1, 2, 3, 4])),
+                ingress_expiry: expiry_time_from_now().as_nanos_since_unix_epoch(),
+            },
+        }
+    }
 
     /// Build a Vec<SignedIngress>.  Convert to IngressPayload and then back to
     /// Vec<SignedIngress>.  Ensure that the two vectors are identical.
     #[test]
     fn into_ingress_payload_and_back() {
-        let ingress_expiry = expiry_time_from_now();
-        let content = HttpCallContent::Call {
-            update: HttpCanisterUpdate {
-                canister_id: Blob(vec![42; 8]),
-                method_name: "some_method".to_string(),
-                arg: Blob(b"".to_vec()),
-                sender: Blob(vec![0x05]),
-                nonce: Some(Blob(vec![1, 2, 3, 4])),
-                ingress_expiry: ingress_expiry.as_nanos_since_unix_epoch(),
-            },
-        };
         let update_messages = vec![
             HttpRequestEnvelope::<HttpCallContent> {
-                content: content.clone(),
+                content: fake_http_call_content("1"),
                 sender_pubkey: Some(Blob(vec![2; 32])),
                 sender_sig: Some(Blob(vec![1; 32])),
                 sender_delegation: None,
             },
             HttpRequestEnvelope::<HttpCallContent> {
-                content: content.clone(),
+                content: fake_http_call_content("2"),
                 sender_pubkey: None,
                 sender_sig: None,
                 sender_delegation: None,
             },
             HttpRequestEnvelope::<HttpCallContent> {
-                content,
+                content: fake_http_call_content("3"),
                 sender_pubkey: Some(Blob(vec![2; 32])),
                 sender_sig: Some(Blob(vec![1; 32])),
                 sender_delegation: Some(vec![SignedDelegation::new(
-                    Delegation::new(vec![1, 2], ingress_expiry),
+                    Delegation::new(vec![1, 2], expiry_time_from_now()),
                     vec![3, 4],
                 )]),
             },
@@ -250,5 +254,70 @@ mod tests {
         let ingress_payload = IngressPayload::from(signed_ingresses.clone());
         let signed_ingresses1 = Vec::<SignedIngress>::try_from(ingress_payload).unwrap();
         assert_eq!(signed_ingresses, signed_ingresses1);
+    }
+
+    #[test]
+    fn test_ingress_payload_deserialization() {
+        // serialization/deserialization of empty payload.
+        let payload = IngressPayload::default();
+        let bytes = bincode::serialize(&payload).unwrap();
+        assert_eq!(
+            bincode::deserialize::<IngressPayload>(&bytes).unwrap(),
+            payload
+        );
+
+        let fake_ingress_message = |method_name| {
+            let message = HttpRequestEnvelope::<HttpCallContent> {
+                content: fake_http_call_content(method_name),
+                sender_pubkey: None,
+                sender_sig: None,
+                sender_delegation: None,
+            };
+
+            SignedIngress::try_from(message).unwrap()
+        };
+
+        // Some test messages.
+        let m1 = fake_ingress_message("m1");
+        let m1_id = m1.id();
+        let m2 = fake_ingress_message("m2");
+        let m3 = fake_ingress_message("m3");
+
+        let msgs = vec![m1, m2, m3];
+        let payload = IngressPayload::from(msgs.clone());
+        // Serialization/deserialization works.
+        let mut bytes = bincode::serialize(&payload).unwrap();
+        assert_eq!(
+            bincode::deserialize::<IngressPayload>(&bytes).unwrap(),
+            payload
+        );
+        // Individual lookup works.
+        assert_matches!(payload.get(0).unwrap(), (_, msg) if msg == msgs[0]);
+        assert_matches!(payload.get(1).unwrap(), (_, msg) if msg == msgs[1]);
+        assert_matches!(payload.get(2).unwrap(), (_, msg) if msg == msgs[2]);
+        // Test IndexOutOfBound.
+        assert_matches!(payload.get(3), Err(IngressPayloadError::IndexOutOfBound(3)));
+        // Converting back to messages should match original
+        assert_eq!(msgs, <Vec<SignedIngress>>::try_from(payload).unwrap());
+
+        // A sub-sequence search function
+        fn find(array: &[u8], subseq: &[u8]) -> Option<usize> {
+            (0..array.len() - subseq.len() + 1).find(|&i| array[i..i + subseq.len()] == subseq[..])
+        }
+
+        // Mutate some byte, deserialization works, but casting back to messages fail.
+        let pos = find(&bytes, m1_id.as_bytes()).unwrap();
+        // `+= 1` may overflow in debug mode.
+        bytes[pos] ^= 1;
+        let payload = bincode::deserialize::<IngressPayload>(&bytes);
+        assert!(payload.is_ok());
+        let payload = payload.unwrap();
+        // get(0) should return error.
+        assert_matches!(
+            payload.get(0),
+            Err(IngressPayloadError::MismatchedMessageIdAtIndex(0))
+        );
+        // Conversion should also fail.
+        assert!(<Vec<_>>::try_from(payload).is_err());
     }
 }
